pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "koopman_learning_tropical_cyclone"
version = "0.0.1"
description = "Here we explore the possibilities to employ the Koopman operator framework to model, simulate and predict the movement of tropical cyclones."
requires-python = "==3.11.9"
authors = [{name = "Eric Brunner", email = "nephts@mailbox.org"}]
dependencies = [
    "ipykernel",
    "ipython",
    "matplotlib",
    "matplotlib-inline",
    "numpy",
    "pandas",
    "scipy",
    "climada",
    "kooplearn",
    "dask",
    "cfgrib",
    "ml_confs",
    "torch",
    "fiona",
    "dask[dataframe]",
    "absl-py",
    "py7zr",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pyright",
    "black",
]
tests = [
  "pytest",
  "pytest-lazy-fixtures",
]

[tool.setuptools]
packages = ["klearn_tcyclone"]

[tool.ruff]
# Default configutations: https://docs.astral.sh/ruff/configuration/
line-length = 88
lint.select = ["E", "F", "B", "RUF", "PT", "UP", "C4", "D"]
lint.ignore = ["RUF002"]
extend-exclude = ["**/*ipynb", "tests/conftest.py", ".venv/**", "klearn_tcyclone/_types/_type_stubs/**"]
target-version = "py311"

[tool.ruff.lint.pydocstyle]
convention = "google"

